Add Disambiguation and DisambiguationUtils (#219)
* initial commit

* rename restoreCheckpoint -> restoreFromCheckpoint

* removed dead code

* initial commit

* final commit
