+++ src/parser/disambiguation/disambiguation.ts
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export const enum ParenthesisDisambiguation {
+    FunctionExpression = "FunctionExpression",
+    ParenthesizedExpression = "ParenthesizedExpression",
+}
+
+export const enum BracketDisambiguation {
+    FieldProjection = "FieldProjection",
+    FieldSelection = "FieldSelection",
+    Record = "Record",
+}
