+++ src/parser/disambiguation/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import * as Disambiguation from "./disambiguation";
+import * as DisambiguationUtils from "./disambiguationUtils";
+
+export { Disambiguation, DisambiguationUtils };
