+++ src/parser/parsers/naive.ts
@@ -14,7 +14,8 @@ import {
 } from "../../common";
 import { Ast, AstUtils, Constant, ConstantUtils, Token } from "../../language";
 import { LexerSnapshot } from "../../lexer";
-import { BracketDisambiguation, IParser, IParserStateCheckpoint, ParenthesisDisambiguation } from "../IParser";
+import { Disambiguation, DisambiguationUtils } from "../disambiguation";
+import { IParser, IParserStateCheckpoint } from "../IParser";
 import { IParserState, IParserStateUtils } from "../IParserState";
 import { NodeIdMapUtils } from "../nodeIdMap";
 
@@ -359,25 +360,7 @@ export function readExpression<S extends IParserState = IParserState>(state: S, 
             return parser.readErrorHandlingExpression(state, parser);
 
         case Token.TokenKind.LeftParenthesis:
-            const triedDisambiguation: Result<
-                ParenthesisDisambiguation,
-                ParseError.UnterminatedSequence
-            > = parser.disambiguateParenthesis(state, parser);
-            if (ResultUtils.isErr(triedDisambiguation)) {
-                throw triedDisambiguation.error;
-            }
-            const disambiguation: ParenthesisDisambiguation = triedDisambiguation.value;
-
-            switch (disambiguation) {
-                case ParenthesisDisambiguation.FunctionExpression:
-                    return parser.readFunctionExpression(state, parser);
-
-                case ParenthesisDisambiguation.ParenthesizedExpression:
-                    return parser.readNullCoalescingExpression(state, parser);
-
-                default:
-                    throw Assert.isNever(disambiguation);
-            }
+            return DisambiguationUtils.readAmbiguousParenthesis(state, parser);
 
         default:
             return parser.readNullCoalescingExpression(state, parser);
@@ -672,10 +655,10 @@ export function readPrimaryExpression<S extends IParserState = IParserState>(
                 break;
 
             case Token.TokenKind.LeftBracket:
-                primaryExpression = readBracketDisambiguation(state, parser, [
-                    BracketDisambiguation.FieldProjection,
-                    BracketDisambiguation.FieldSelection,
-                    BracketDisambiguation.Record,
+                primaryExpression = DisambiguationUtils.readAmbiguousBracket(state, parser, [
+                    Disambiguation.BracketDisambiguation.FieldProjection,
+                    Disambiguation.BracketDisambiguation.FieldSelection,
+                    Disambiguation.BracketDisambiguation.Record,
                 ]);
                 break;
 
@@ -779,10 +762,14 @@ export function readRecursivePrimaryExpression<S extends IParserState = IParserS
         } else if (maybeCurrentTokenKind === Token.TokenKind.LeftBrace) {
             recursiveExpressions.push(parser.readItemAccessExpression(state, parser));
         } else if (maybeCurrentTokenKind === Token.TokenKind.LeftBracket) {
-            const bracketExpression: Ast.TFieldAccessExpression = readBracketDisambiguation(state, parser, [
-                BracketDisambiguation.FieldProjection,
-                BracketDisambiguation.FieldSelection,
-            ]) as Ast.TFieldAccessExpression;
+            const bracketExpression: Ast.TFieldAccessExpression = DisambiguationUtils.readAmbiguousBracket(
+                state,
+                parser,
+                [
+                    Disambiguation.BracketDisambiguation.FieldProjection,
+                    Disambiguation.BracketDisambiguation.FieldSelection,
+                ],
+            ) as Ast.TFieldAccessExpression;
             recursiveExpressions.push(bracketExpression);
         } else {
             continueReadingValues = false;
@@ -1945,127 +1932,6 @@ function tryReadPrimitiveType<S extends IParserState = IParserState>(
     return ResultUtils.okFactory(astNode);
 }
 
-// ------------------------------------
-// ---------- Disambiguation ----------
-// ------------------------------------
-
-export function disambiguateParenthesis<S extends IParserState = IParserState>(
-    state: S,
-    parser: IParser<S>,
-): Result<ParenthesisDisambiguation, ParseError.UnterminatedSequence> {
-    state.maybeCancellationToken?.throwIfCancelled();
-
-    const initialTokenIndex: number = state.tokenIndex;
-    const tokens: ReadonlyArray<Token.Token> = state.lexerSnapshot.tokens;
-    const totalTokens: number = tokens.length;
-    let nestedDepth: number = 1;
-    let offsetTokenIndex: number = initialTokenIndex + 1;
-
-    while (offsetTokenIndex < totalTokens) {
-        const offsetTokenKind: Token.TokenKind = tokens[offsetTokenIndex].kind;
-
-        if (offsetTokenKind === Token.TokenKind.LeftParenthesis) {
-            nestedDepth += 1;
-        } else if (offsetTokenKind === Token.TokenKind.RightParenthesis) {
-            nestedDepth -= 1;
-        }
-
-        if (nestedDepth === 0) {
-            // '(x as number) as number' could either be either case,
-            // so we need to consume test if the trailing 'as number' is followed by a FatArrow.
-            if (IParserStateUtils.isTokenKind(state, Token.TokenKind.KeywordAs, offsetTokenIndex + 1)) {
-                const checkpoint: IParserStateCheckpoint = parser.createCheckpoint(state);
-                unsafeMoveTo(state, offsetTokenIndex + 2);
-
-                try {
-                    parser.readNullablePrimitiveType(state, parser);
-                } catch {
-                    parser.restoreFromCheckpoint(state, checkpoint);
-                    if (IParserStateUtils.isOnTokenKind(state, Token.TokenKind.FatArrow)) {
-                        return ResultUtils.okFactory(ParenthesisDisambiguation.FunctionExpression);
-                    } else {
-                        return ResultUtils.okFactory(ParenthesisDisambiguation.ParenthesizedExpression);
-                    }
-                }
-
-                let disambiguation: ParenthesisDisambiguation;
-                if (IParserStateUtils.isOnTokenKind(state, Token.TokenKind.FatArrow)) {
-                    disambiguation = ParenthesisDisambiguation.FunctionExpression;
-                } else {
-                    disambiguation = ParenthesisDisambiguation.ParenthesizedExpression;
-                }
-
-                parser.restoreFromCheckpoint(state, checkpoint);
-                return ResultUtils.okFactory(disambiguation);
-            } else {
-                if (IParserStateUtils.isTokenKind(state, Token.TokenKind.FatArrow, offsetTokenIndex + 1)) {
-                    return ResultUtils.okFactory(ParenthesisDisambiguation.FunctionExpression);
-                } else {
-                    return ResultUtils.okFactory(ParenthesisDisambiguation.ParenthesizedExpression);
-                }
-            }
-        }
-
-        offsetTokenIndex += 1;
-    }
-
-    return ResultUtils.errFactory(IParserStateUtils.unterminatedParenthesesError(state));
-}
-
-// WARNING: Only updates tokenIndex and currentTokenKind,
-//          Manual management of TokenRangeStack is assumed.
-//          Best used in conjunction with backup/restore using ParserState.
-function unsafeMoveTo(state: IParserState, tokenIndex: number): void {
-    const tokens: ReadonlyArray<Token.Token> = state.lexerSnapshot.tokens;
-    state.tokenIndex = tokenIndex;
-
-    if (tokenIndex < tokens.length) {
-        state.maybeCurrentToken = tokens[tokenIndex];
-        state.maybeCurrentTokenKind = state.maybeCurrentToken.kind;
-    } else {
-        state.maybeCurrentToken = undefined;
-        state.maybeCurrentTokenKind = undefined;
-    }
-}
-
-export function disambiguateBracket<S extends IParserState = IParserState>(
-    state: S,
-    _parser: IParser<S>,
-): Result<BracketDisambiguation, ParseError.UnterminatedSequence> {
-    state.maybeCancellationToken?.throwIfCancelled();
-
-    const tokens: ReadonlyArray<Token.Token> = state.lexerSnapshot.tokens;
-    let offsetTokenIndex: number = state.tokenIndex + 1;
-    const offsetToken: Token.Token = tokens[offsetTokenIndex];
-
-    if (!offsetToken) {
-        return ResultUtils.errFactory(IParserStateUtils.unterminatedBracketError(state));
-    }
-
-    let offsetTokenKind: Token.TokenKind = offsetToken.kind;
-    if (offsetTokenKind === Token.TokenKind.LeftBracket) {
-        return ResultUtils.okFactory(BracketDisambiguation.FieldProjection);
-    } else if (offsetTokenKind === Token.TokenKind.RightBracket) {
-        return ResultUtils.okFactory(BracketDisambiguation.Record);
-    } else {
-        const totalTokens: number = tokens.length;
-        offsetTokenIndex += 1;
-        while (offsetTokenIndex < totalTokens) {
-            offsetTokenKind = tokens[offsetTokenIndex].kind;
-
-            if (offsetTokenKind === Token.TokenKind.Equal) {
-                return ResultUtils.okFactory(BracketDisambiguation.Record);
-            } else if (offsetTokenKind === Token.TokenKind.RightBracket) {
-                return ResultUtils.okFactory(BracketDisambiguation.FieldSelection);
-            }
-
-            offsetTokenIndex += 1;
-        }
-
-        return ResultUtils.errFactory(IParserStateUtils.unterminatedBracketError(state));
-    }
-}
-
 // -------------------------------------
 // ---------- key-value pairs ----------
 // -------------------------------------
@@ -2653,42 +2519,6 @@ function maybeReadLiteralAttributes<S extends IParserState = IParserState>(
     } else {
         IParserStateUtils.incrementAttributeCounter(state);
         return undefined;
-    }
-}
-
-// -------------------------------------------------------
-// ---------- Helper functions (disambiguation) ----------
-// -------------------------------------------------------
-
-export function readBracketDisambiguation<S extends IParserState = IParserState>(
-    state: S,
-    parser: IParser<S>,
-    allowedVariants: ReadonlyArray<BracketDisambiguation>,
-): Ast.FieldProjection | Ast.FieldSelector | Ast.RecordExpression {
-    state.maybeCancellationToken?.throwIfCancelled();
-
-    const triedDisambiguation: Result<
-        BracketDisambiguation,
-        ParseError.UnterminatedSequence
-    > = parser.disambiguateBracket(state, parser);
-    if (ResultUtils.isErr(triedDisambiguation)) {
-        throw triedDisambiguation.error;
-    }
-    const disambiguation: BracketDisambiguation = triedDisambiguation.value;
-    ArrayUtils.assertIn(allowedVariants, disambiguation, `invalid disambiguation`);
-
-    switch (disambiguation) {
-        case BracketDisambiguation.FieldProjection:
-            return parser.readFieldProjection(state, parser);
-
-        case BracketDisambiguation.FieldSelection:
-            return parser.readFieldSelection(state, parser);
-
-        case BracketDisambiguation.Record:
-            return parser.readRecordExpression(state, parser);
-
-        default:
-            throw Assert.isNever(disambiguation);
     }
 }
